feat: add ResponsePanel component for displaying gRPC responses
feat: update Button component styles and variants

feat: enhance Card component styles for better theming

feat: improve Checkbox component with indeterminate state support

feat: refactor Input component styles for consistency

feat: update Label component styles for better theming

feat: enhance Select component styles for improved UX

feat: refactor Textarea component styles for consistency

feat: add ThemeToggle component for switching themes

feat: implement ThemeContext for managing theme state

style: update global CSS for light and dark themes

fix: update i18n imports to use absolute paths

feat: add optional translations in locales

refactor: remove unused presets state management

refactor: clean up request state management

chore: remove obsolete test script

chore: update TypeScript configuration for path aliases

chore: configure Vite to resolve path aliases
