+++ package.json
@@ -5,14 +5,12 @@
   "type": "module",
   "packageManager": "pnpm@9.0.0",
   "scripts": {
-    "dev": "./src-tauri/target/release/grpc-bridge",
+    "dev": "pnpm tauri dev",
+    "dev:frontend": "vite dev",
     "build": "pnpm vite build",
-    "preview": "pnpm vite preview",
-    "tauri:dev": "pnpm tauri dev",
-    "tauri:build": "pnpm tauri build",
-    "lint": "pnpm eslint .",
     "cross-build": "./cross-build.sh",
-    "setup": "pnpm install && rustup target add x86_64-pc-windows-gnu x86_64-apple-darwin",
+    "lint": "pnpm eslint .",
+    "setup": "./setup-dev.sh",
     "clean": "rm -rf dist/ src-tauri/target/ node_modules/",
     "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,css,md}\"",
     "type-check": "tsc --noEmit",
@@ -23,38 +21,21 @@
     "release:major": "pnpm version major && git push && git push --tags"
   },
   "dependencies": {
-    "@monaco-editor/react": "^4.6.0",
-    "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-checkbox": "^1.3.3",
-    "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/react-label": "^2.1.7",
-    "@radix-ui/react-popover": "^1.1.15",
-    "@radix-ui/react-scroll-area": "^1.2.10",
     "@radix-ui/react-select": "^2.2.6",
-    "@radix-ui/react-separator": "^1.1.7",
     "@radix-ui/react-slot": "^1.2.3",
-    "@radix-ui/react-switch": "^1.2.6",
-    "@radix-ui/react-toast": "^1.2.15",
     "@tanstack/react-query": "^5.50.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "framer-motion": "^11.0.0",
-    "fuse.js": "^7.0.0",
     "i18next": "^25.5.2",
     "i18next-browser-languagedetector": "^8.2.0",
-    "jsonrepair": "^3.0.0",
-    "kbar": "0.1.0-beta.48",
     "lucide-react": "^0.378.0",
-    "rc-tree": "^5.13.1",
     "react": "^18.3.0",
     "react-dom": "^18.3.0",
-    "react-error-boundary": "^4.0.13",
-    "react-hook-form": "^7.52.0",
     "react-hot-toast": "^2.4.1",
     "react-i18next": "^15.7.3",
-    "react-virtuoso": "^4.10.0",
     "tailwind-merge": "^3.3.1",
-    "zod": "^3.23.0",
     "zustand": "^4.5.0"
   },
   "devDependencies": {
