+++ src/components/ui/ThemeToggle.tsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import { Sun, Moon } from 'lucide-react';
+import { useTheme } from '@/context/ThemeContext';
+import { Button } from './Button';
+
+export const ThemeToggle: React.FC = () => {
+  const { theme, toggleTheme } = useTheme();
+
+  const handleToggle = () => {
+    console.log('Theme toggle clicked, current theme:', theme);
+    toggleTheme();
+  };
+
+  return (
+    <Button
+      variant="ghost"
+      size="icon"
+      onClick={handleToggle}
+      className="h-8 w-8 hover:bg-accent hover:text-accent-foreground"
+      title={theme === 'light' ? 'Switch to dark mode' : 'Switch to light mode'}
+    >
+      {theme === 'light' ? (
+        <Moon className="h-4 w-4" />
+      ) : (
+        <Sun className="h-4 w-4" />
+      )}
+    </Button>
+  );
+};
