+++ src/context/ThemeContext.tsx
@@ -1,0 +1,68 @@
+import React, { createContext, useContext, useEffect, useState } from 'react';
+
+type Theme = 'light' | 'dark';
+
+interface ThemeContextType {
+  theme: Theme;
+  toggleTheme: () => void;
+  setTheme: (theme: Theme) => void;
+}
+
+const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+};
+
+interface ThemeProviderProps {
+  children: React.ReactNode;
+}
+
+export const ThemeProvider: React.FC<ThemeProviderProps> = ({ children }) => {
+  const [theme, setThemeState] = useState<Theme>(() => {
+    // Check if we're in browser environment
+    if (typeof window === 'undefined') return 'light';
+
+    // Check localStorage first, then system preference
+    const stored = localStorage.getItem('theme') as Theme;
+    if (stored && (stored === 'light' || stored === 'dark')) {
+      return stored;
+    }
+    return window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light';
+  });
+
+  useEffect(() => {
+    if (typeof window === 'undefined') return;
+
+    const root = document.documentElement;
+
+    // Remove existing theme classes
+    root.classList.remove('light', 'dark');
+
+    // Add current theme class
+    root.classList.add(theme);
+
+    // Store in localStorage
+    localStorage.setItem('theme', theme);
+
+    console.log('Theme changed to:', theme, 'Classes:', root.classList.toString());
+  }, [theme]);
+
+  const toggleTheme = () => {
+    setThemeState(prev => prev === 'light' ? 'dark' : 'light');
+  };
+
+  const setTheme = (newTheme: Theme) => {
+    setThemeState(newTheme);
+  };
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme, setTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
