+++ src/i18n.ts
@@ -3,9 +3,9 @@ import { initReactI18next } from 'react-i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 
 // Import translation files
-import en from './locales/en.json';
-import ja from './locales/ja.json';
-import ko from './locales/ko.json';
+import en from '@/locales/en.json';
+import ja from '@/locales/ja.json';
+import ko from '@/locales/ko.json';
 
 const resources = {
   en: {
