+++ src/locales/ja.json
@@ -22,7 +22,8 @@
         "format": "フォーマット",
         "showDiff": "差分を表示",
         "hideDiff": "差分を隠す",
-        "warning": "警告"
+        "warning": "警告",
+        "optional": "オプション"
     },
     "app": {
         "title": "gRPC Bridge",
@@ -41,7 +42,10 @@
         "fileCount_plural": "{{count}}個のファイル",
         "protoRootPath": "Protoルートパス",
         "pathPlaceholder": "/abs/path/to/protos",
-        "selectRoot": "ルートを選択"
+        "selectRoot": "ルートを選択",
+        "confirmRemove": "このルートを削除しますか？",
+        "removeSuccess": "ルートが正常に削除されました",
+        "removeRoot": "ルートを削除"
     },
     "grpc": {
         "configuration": "設定",
@@ -70,13 +74,11 @@
         "requestFailed": "リクエストが失敗しました",
         "skeleton": "スケルトン",
         "noPreviousPayload": "以前のペイロードがありません",
-        "savePreset": "プリセットを保存",
         "payloadPlaceholder": "{\"key\": \"value\"}",
-        "loadPreset": "プリセットを読み込み",
-        "renamePreset": "プリセット名を変更",
         "authToken": "認証トークン",
         "authTokenPlaceholder": "Bearerトークン",
-        "autoAuth": "Authorizationヘッダーを自動追加"
+        "autoAuth": "Authorizationヘッダーを自動追加",
+        "executeRequest": "リクエストテスト"
     },
     "menu": {
         "file": "ファイル",
