+++ src/main.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { Toaster } from 'react-hot-toast';
-import App from './App';
-import './global.css';
+import App from '@/App';
+import '@/global.css';
 
 const qc = new QueryClient();
